test: add integration tests for pyproject.toml handling
