+++ crates/pixi-build-python/tests/integration/main.rs
@@ -1,0 +1,1 @@
+mod pyproject;
